old system/agents/application-processing-agent/pyproject.toml
[project]
name = "application-processing-agent"
version = "0.1.0"
description = "Application Processing Agent for extracting and validating medical application data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]==0.3.10",
    
    "starlette==0.49.3",
    "uvicorn==0.38.0",
    "sse-starlette==3.0.3",
    
    "langchain-core==1.0.2",
    "langchain-openai==1.0.1",
    "langchain-community==0.4.1",
    "langgraph==1.0.2",
    
    "pydantic==2.12.3",
    "pydantic-settings==2.11.0",
    
    "python-dotenv==1.2.1",
    "click==8.3.0",
    
    "httpx==0.28.1",
    
    "anyio==4.11.0",
]